refactor: rename package, convert to mapper
